// src/logic/types/GastoCategoria.ts
import { Ionicons } from '@expo/vector-icons';

// Enum de categorías de gasto (solo claves)
export enum GastoCategoria {
  ALIMENTACION_Y_BEBIDAS = 'ALIMENTACION_Y_BEBIDAS',
  TRANSPORTE_Y_VEHICULOS = 'TRANSPORTE_Y_VEHICULOS',
  VIVIENDA_Y_SERVICIOS = 'VIVIENDA_Y_SERVICIOS',
  SALUD_Y_BIENESTAR = 'SALUD_Y_BIENESTAR',
  EDUCACION_Y_CULTURA = 'EDUCACION_Y_CULTURA',
  COMPRAS_EN_LINEA = 'COMPRAS_EN_LINEA',
  FINANZAS = 'FINANZAS',
  FAMILIA_Y_CUIDADO = 'FAMILIA_Y_CUIDADO',
  OCIO_Y_ENTRETENIMIENTO = 'OCIO_Y_ENTRETENIMIENTO',
  IMPUESTOS_Y_TRIBUTOS = 'IMPUESTOS_Y_TRIBUTOS',
  AHORRO_E_INVERSION = 'AHORRO_E_INVERSION',
  ROPA_Y_ACCESORIOS = 'ROPA_Y_ACCESORIOS',
  REGALOS_Y_DONACIONES = 'REGALOS_Y_DONACIONES',
  MANTENIMIENTO_Y_CUIDADO = 'MANTENIMIENTO_Y_CUIDADO',
  VIAJES_Y_VACACIONES = 'VIAJES_Y_VACACIONES',
  SUSCRIPCIONES_Y_SERVICIOS = 'SUSCRIPCIONES_Y_SERVICIOS',
  CUIDADO_PERSONAL_Y_DEPORTES = 'CUIDADO_PERSONAL_Y_DEPORTES',
  MASCOTAS_Y_ANIMALES = 'MASCOTAS_Y_ANIMALES',
  ARTE_Y_CULTURA = 'ARTE_Y_CULTURA',
  CARIDAD_Y_VOLUNTARIADO = 'CARIDAD_Y_VOLUNTARIADO',
  COMUNICACIONES = 'COMUNICACIONES',
  EDUCACION_FINANCIERA = 'EDUCACION_FINANCIERA',
  INVERSIONES_ALTERNATIVAS = 'INVERSIONES_ALTERNATIVAS',
  FONDO_DE_EMERGENCIA = 'FONDO_DE_EMERGENCIA',
}

export const GastoCategoriaInfo = {
  [GastoCategoria.ALIMENTACION_Y_BEBIDAS]: { name: 'Alimentación y Bebidas', icon: 'restaurant' },
  [GastoCategoria.TRANSPORTE_Y_VEHICULOS]: { name: 'Transporte y Vehículos', icon: 'car' },
  [GastoCategoria.VIVIENDA_Y_SERVICIOS]: { name: 'Vivienda y Servicios', icon: 'home' },
  [GastoCategoria.SALUD_Y_BIENESTAR]: { name: 'Salud y Bienestar', icon: 'medkit' },
  [GastoCategoria.EDUCACION_Y_CULTURA]: { name: 'Educación y Cultura', icon: 'school' },
  [GastoCategoria.COMPRAS_EN_LINEA]: { name: 'Compras en Línea', icon: 'cart' },
  [GastoCategoria.FINANZAS]: { name: 'Finanzas', icon: 'wallet' },
  [GastoCategoria.FAMILIA_Y_CUIDADO]: { name: 'Familia y Cuidado', icon: 'person' },
  [GastoCategoria.OCIO_Y_ENTRETENIMIENTO]: { name: 'Ocio y Entretenimiento', icon: 'ios-film' },
  [GastoCategoria.IMPUESTOS_Y_TRIBUTOS]: { name: 'Impuestos y Tributos', icon: 'cash' },
  [GastoCategoria.AHORRO_E_INVERSION]: { name: 'Ahorro e Inversión', icon: 'bar-chart' },
  [GastoCategoria.ROPA_Y_ACCESORIOS]: { name: 'Ropa y Accesorios', icon: 'shirt' },
  [GastoCategoria.REGALOS_Y_DONACIONES]: { name: 'Regalos y Donaciones', icon: 'gift' },
  [GastoCategoria.MANTENIMIENTO_Y_CUIDADO]: { name: 'Mantenimiento y Cuidado', icon: 'hammer' },
  [GastoCategoria.VIAJES_Y_VACACIONES]: { name: 'Viajes y Vacaciones', icon: 'airplane' },
  [GastoCategoria.SUSCRIPCIONES_Y_SERVICIOS]: { name: 'Suscripciones y Otros Servicios', icon: 'notifications' },
  [GastoCategoria.CUIDADO_PERSONAL_Y_DEPORTES]: { name: 'Cuidado Personal y Deportes', icon: 'fitness' },
  [GastoCategoria.MASCOTAS_Y_ANIMALES]: { name: 'Mascotas y Animales', icon: 'paw' },
  [GastoCategoria.ARTE_Y_CULTURA]: { name: 'Arte y Cultura', icon: 'brush' },
  [GastoCategoria.CARIDAD_Y_VOLUNTARIADO]: { name: 'Caridad y Voluntariado', icon: 'heart' },
  [GastoCategoria.COMUNICACIONES]: { name: 'Comunicaciones', icon: 'chatbubbles' },
  [GastoCategoria.EDUCACION_FINANCIERA]: { name: 'Educación Financiera', icon: 'book' },
  [GastoCategoria.INVERSIONES_ALTERNATIVAS]: { name: 'Inversiones Alternativas', icon: 'globe' },
  [GastoCategoria.FONDO_DE_EMERGENCIA]: { name: 'Fondo de Emergencia', icon: 'shield' },
}